+++ lib/rails_route_tester/railtie.rb
@@ -8,6 +8,7 @@ module RailsRouteTester
       load "rails_route_tester/rake_tasks/routes.rake"
       load "rails_route_tester/rake_tasks/pom.rake"
       load "rails_route_tester/rake_tasks/tests.rake"
+      load "rails_route_tester/rake_tasks/results.rake"
     end
   end
 end